+++ Localization/deDE.Plugins.lua
@@ -5,7 +5,7 @@ if not L then return end
 -- [[ https://legacy.curseforge.com/wow/addons/krowi-achievement-filter/localization ]] --
 -- [[ Everything after this line is automatically generated from CurseForge and is not meant for manual edit - SOURCETOKEN - AUTOGENTOKEN ]] --
 
--- [[ Exported at 2023-09-17 08-51-06 ]] --
+-- [[ Exported at 2024-01-21 12-58-50 ]] --
 L["AchievementsTabFixes"] = "\"Erfolge\" Tab Korrekturen"
 L["AchievementsTabFixes Desc"] = "Dieses Plugin bietet einige Korrekturen für den Tab \"Erfolge\", die unten umgeschaltet werden können."
 L["Add To Right Click Menu"] = "Zum Rechtsklick-Menü hinzufügen "