+++ Localization/frFR.Plugins.lua
@@ -4,3 +4,10 @@ if not L then return end
 
 -- [[ https://legacy.curseforge.com/wow/addons/krowi-achievement-filter/localization ]] --
 -- [[ Everything after this line is automatically generated from CurseForge and is not meant for manual edit - SOURCETOKEN - AUTOGENTOKEN ]] --
+
+-- [[ Exported at 2024-01-21 12-58-53 ]] --
+L["At least version is required"] = "La version minimum  '{version}' est requise."
+L["ElvUI"] = true
+L["GW2_UI"] = "GW 2 UI"
+L["IAT Tactics"] = true
+L["Instance Achievement Tracker"] = true